backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.5
networkx==3.4.2
pandas==2.2.3
pydantic==2.10.4
sqlparse==0.5.3
